refactor: 💡 download

+++ deploy/src/main.rs
@@ -11,6 +11,7 @@ use std::fs::File;
 use std::io::stdin;
 use std::process::Command;
 use std::{fs, io};
+use std::path::{Path, PathBuf};
 use tokio::io::AsyncWriteExt;
 
 #[tokio::main]
@@ -20,54 +21,70 @@ async fn main() -> Result<()> {
 
     if args.release.is_some() {
         config.set_version(args.release.unwrap().as_str());
+    } else {
+        // get latest version
+        let latest_version = reqwest::get("https://data.jsdelivr.com/v1/package/gh/ZingerLittleBee/server_bee-backend")
+            .await?
+            .json::<serde_json::Value>()
+            .await?;
+
+        if let Some(version_value) = latest_version.get("versions") {
+            if let Some(version_vec) = version_value.as_array() {
+                if let Some(version) = version_vec.first() {
+                    if let Some(version_str) = version.as_str() {
+                        println!("latest version: {}", version_str);
+                        config.set_version(version_str);
+                    }
+                }
+            }
+        }
     }
 
     if args.port.is_some() {
         config.set_port(Port::new(args.port.unwrap()));
     }
 
-    // get latest version
-    // let latest_version = reqwest::get("https://data.jsdelivr.com/v1/package/gh/ZingerLittleBee/server_bee-backend")
-    //     .await?
-    //     .json::<serde_json::Value>()
-    //     .await?;
-    //
-    // println!("Latest version: {:#?}", latest_version.get("versions").unwrap().as_array().unwrap().first().unwrap().as_str().unwrap());
-    //
-    // return Ok(());
+    if !Path::new(&config.web_bin_path()).exists() {
+        let mut response = reqwest::Client::new()
+            .get(config.bin_zip_url().to_str().unwrap())
+            .send()
+            .await?;
+        println!("正在下载 {}", config.bin_zip_url().display());
 
-    let mut response = reqwest::Client::new()
-        .get(config.bin_zip_url().to_str().unwrap())
-        .send()
-        .await?;
-    println!("正在下载 {}", config.bin_zip_url().display());
+        if response.status().as_u16() >= 400 {
+            println!("文件下载失败 {}", response.status());
+            return Ok(());
+        }
 
-    if response.status().as_u16() >= 400 {
-        println!("文件下载失败 {}", response.status());
-        return Ok(());
+        let web_file_path = config.web_bin_zip_path();
+
+        if let Some(p) = web_file_path.parent() {
+            if !p.exists() {
+                fs::create_dir_all(&p).unwrap();
+            }
+        }
+
+        println!("文件存放路径: {}", web_file_path.display());
+
+        let mut file = tokio::fs::File::create(&web_file_path).await?;
+
+        while let Some(mut item) = response.chunk().await? {
+            file.write_all_buf(item.borrow_mut()).await?;
+        }
+        println!("文件 {} 下载成功, 正在解压", web_file_path.display());
+
+        let tokio_file = tokio::fs::File::open(web_file_path).await?;
+        let std_file = tokio_file.into_std().await;
+
+        unzip(std_file, config.web_bin_dir());
+        println!("文件解压完毕");
+    } else {
+        println!("文件已存在, 跳过下载");
     }
-
-    let web_file_path = config.web_bin_zip_path();
-    println!("下载文件路径: {}", web_file_path.display());
-
-    let mut file = tokio::fs::File::create(&web_file_path).await?;
-
-    while let Some(mut item) = response.chunk().await? {
-        file.write_all_buf(item.borrow_mut()).await?;
-    }
-    println!("文件 {} 下载成功, 正在解压", web_file_path.display());
-
-    let tokio_file = tokio::fs::File::open(web_file_path).await?;
-    let std_file = tokio_file.into_std().await;
-
-    unzip(std_file);
-    println!("文件解压完毕");
-
-    return Ok(());
 
     config.set_auto_launch(!args.auto_launch);
 
-    start_process(Config::web_bin_path().to_str().unwrap());
+    start_process(config.web_bin_path().to_str().unwrap(), config.log_path().to_str().unwrap());
 
     println!("按回车键退出...");
     stdin()
@@ -77,15 +94,18 @@ async fn main() -> Result<()> {
     Ok(())
 }
 
-fn unzip(file: File) {
+fn unzip(file: File, out_dir: PathBuf) {
     let mut archive = zip::ZipArchive::new(file).unwrap();
 
     for i in 0..archive.len() {
         let mut file = archive.by_index(i).unwrap();
-        let outpath = match file.enclosed_name() {
+
+        let file_name = match file.enclosed_name() {
             Some(path) => path.to_owned(),
             None => continue,
         };
+
+        let outpath = out_dir.join(file_name);
 
         {
             let comment = file.comment();
@@ -134,7 +154,7 @@ fn start_process(bin_full_path: &str) {
 }
 
 #[cfg(not(windows))]
-fn start_process(bin_full_path: &str) {
+fn start_process(bin_full_path: &str, log_path: &str) {
     use std::time::Duration;
     use std::thread;
 
@@ -143,9 +163,15 @@ fn start_process(bin_full_path: &str) {
         bin_full_path
     );
 
+    println!(
+        "日志路径: {}",
+        log_path
+    );
+
     let cmd = format!(
-        "nohup {} > serverbee-web.log &",
-        bin_full_path
+        "nohup {} > {} &",
+        bin_full_path,
+        log_path
     );
 
     Command::new("sh")
@@ -153,9 +179,9 @@ fn start_process(bin_full_path: &str) {
         .spawn()
         .expect("运行 serverbee-web 失败, 请尝试手动运行");
 
-    let out = Command::new("cat").arg("serverbee-web.log").output();
+    let out = Command::new("cat").arg(log_path).output();
 
     thread::sleep(Duration::from_secs(1));
-    println!("启动日志: {:?}", out.unwrap());
-    println!("完整日志请查看 serverbee-web.log 文件");
+    println!("启动日志: {:?}", String::from_utf8_lossy(&*out.unwrap().stdout));
+    println!("完整日志请查看 {}", log_path);
 }
